third_party/blink/web_tests/external/wpt/scroll-animations/css/animation-duration-auto.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Property animation value 'auto cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim'
  assert_equals: expected "0s cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim" but got "cubic-bezier(0, -2, 1, 3) 0s -3s 4 reverse both paused anim"
Harness: the test ran to completion.
